code/requirements.txt
google-api-python-client==1.5.5
google-cloud-bigquery==0.22.1
google-cloud-pubsub==0.22.0
oauth2client==4.0.0
